VP4Optim

Documentation for VP4Optim.

VP4Optim.Model

VP4Optim.A

VP4Optim.Bb!

VP4Optim.P

VP4Optim.c

VP4Optim.f

VP4Optim.f

VP4Optim.fg!

VP4Optim.fg!

VP4Optim.fgh!

VP4Optim.fgh!

VP4Optim.par

VP4Optim.par!

VP4Optim.par!

VP4Optim.par_changed!

VP4Optim.par_sym

VP4Optim.sym

VP4Optim.val

VP4Optim.x

VP4Optim.x!

VP4Optim.x!

VP4Optim.x_changed!

VP4Optim.x_sym

VP4Optim.y

VP4Optim.y!

VP4Optim.y_model

VP4Optim.χ2

VP4Optim.∂Bb!

VP4Optim.∂∂Bb!

VP4Optim.Model — Type

Model{Ny,Nx,Nc,T}

Abstract supertype of any model specification.

Type parameters

Ny::Int: length of data vector y.

Nx::Int: number of variable parameters.

Nc::Int: number of linear coefficients.

T <: Union{Float64, ComplexF64}: typeof(y)

Recommended use

import VP4Optim as VP

mutable struct SpecificModel{Ny,Nx,Nc,T} <: VP.Model{Ny,Nx,Nc,T}
    # mandatory fields of any Model instance
    sym::Vector{Symbol}         # names of all parameters, variable and fixed 
    x_sym::Vector{Symbol}       # names of variable parameters
    par_sym::Vector{Symbol}     # names of fixed parameters
    val::Vector{Float64}        # values of all parameters, variable and fixed
    x_ind::SVector{Nx,Int}      # indices of variable parameters in field val (order defined by x_sym)
    par_ind::Vector{Int}        # indices of fixed parameters in field val (order defined by par_sym)
    y::SVector{Ny,T}            # actual data vector
    y2::Float64                 # == real(y' * y) (automatically calculated in generic method VP.y!)

    # model-specific information
    # ....
end

VP4Optim.A — Method

A(::Model)

Return VARPRO matrix A.

Must be implemented by each model.

VP4Optim.Bb! — Method

Bb!(mod::Model)

TBW

VP4Optim.P — Method

P(mod::Model{Ny,Nx,Nc,T}, x) where {Ny,Nx,Nc,T}

Returns Hessian of model mod at x.

Can be used as preconditioner.

VP4Optim.c — Method

c(mod::Model)

Return VARPRO vector c.

Calculates generic soluion c = B \ b. Can be replaced by model-specific implementation, if desired (e.g. for performance improvements).

VP4Optim.f — Method

f(y2, B, b)

Should not be called directly.

VP4Optim.f — Method

f(mod::Model)

Return function f of argument x to be minimized, as expected by Optim.jl

VP4Optim.fg! — Method

fg!(mod::Model)

Return function fg! of three arguments (F, G, x) as expected by Optim.jl.

VP4Optim.fg! — Method

fg!(::Model{Ny,Nx,Nc,T}, F, G, y2, B, b, ∂B, ∂b) where {Ny,Nx,Nc,T}

Should not be called directly.

VP4Optim.fgh! — Method

fgh!(mod::Model)

Return function fgh! of four arguments (F, G, H, x) as expected by Optim.jl.

VP4Optim.fgh! — Method

fgh!(::Model{Ny,Nx,Nc,T}, F, G, H, y2, B, b, ∂B, ∂b, ∂∂B, ∂∂b) where {Ny,Nx,Nc,T}

Should not be called directly.

VP4Optim.par! — Method

par!(mod::Model, p_syms::AbstractArray, p_vals::AbstractArray)

TBW

VP4Optim.par! — Method

par!(mod::Model, new_par::AbstractArray)

Reset fixed parameter values.

VP4Optim.par — Method

par(mod::Model)

TBW

VP4Optim.par_changed! — Method

par_changed!(::Model)

Notify model that (some of the) fixed parameters par have changed.

Function is automatically called by Can be used to implement

VP4Optim.par_sym — Method

par_sym(mod::Model)

Returns iterable of fixed model parameters.

VP4Optim.sym — Method

sym(mod::Model)

Return iterable of model parameter names (each with type == Symbol).

All model parameters are returned, variable and fixed ones. These are assumed to be stored in the field sym of mod.

VP4Optim.val — Method

val(mod::Model)

Returns vector (::Vector{Float64}) of all model parameters.

VP4Optim.x! — Method

x!(mod::Model, new_x::AbstractArray)

TBW

VP4Optim.x! — Method

x!(mod::Model, new_x::AbstractArray)

TBW

VP4Optim.x — Method

x(mod::Model)

TBW

VP4Optim.x_changed! — Method

x_changed!(::Model)

TBW

VP4Optim.x_sym — Method

x_sym(mod::Model)

Return iterable of variable model parameters.

VP4Optim.y! — Method

y!(mod::Model{Ny,Nx,Nc,T}, new_y::AbstractArray) where {Ny,Nx,Nc,T}

Set the data vector.

VP4Optim.y — Method

y(mod::Model)

Returns the data vector

VP4Optim.y_model — Method

y_model(mod::Model)

Compute model prediction A(mod) * c(mod) at actual x.

VP4Optim.χ2 — Method

χ2(mod::Model)

Return χ² of actual model.

VP4Optim.∂Bb! — Method

∂Bb!(::Model)

TBW

VP4Optim.∂∂Bb! — Method

∂∂Bb!(::Model)

TBW